requirements.txt
nltk==3.0.0
numpy==1.9.1
oauthlib==0.7.2
python-twitter==2.2
requests==2.5.1
requests-oauthlib==0.4.2
wsgiref==0.1.2
